refactor(a2a): move card/remote-agent properties to boot; card construction in AgentCards
P2: AgentCardProperties and RemoteAgentProperties relocated from
engine.a2a to boot (flat, per bootIsFlat ArchUnit rule).
A2aClientAutoConfiguration updated to import boot.RemoteAgentProperties.
RuntimeAutoConfiguration @EnableConfigurationProperties already in boot.

P3: createAgentCard(String) removed from AgentCardProperties; logic
moved into a new AgentCards.create(name,description,version,endpoint,
organization,organizationUrl) overload that applies the same blank-to-
default behavior and optional provider override. AgentCardProperties in
boot now holds only fields, getters/setters, and hasExplicitName() with
no A2A SDK imports. RuntimeAutoConfiguration.a2aAgentCard() updated to
call the new 6-param AgentCards.create() — name-resolution logic
(explicit-name → default-agent-id → first handler) unchanged.
